Adds the Service Layer

+++ LeMaitreBusiness/src/br/cefetmg/LeMaitre/model/dao/EmployeeDAO.java
@@ -13,7 +13,7 @@ import br.cefetmg.LeMaitre.model.exception.PersistenceException;
  * @author Thalesgsn
  */
 public interface EmployeeDAO {
-    public Long insert(Employee employee) throws PersistenceException;
+    public Integer insert(Employee employee) throws PersistenceException;
     public boolean update(Employee employee) throws PersistenceException;
     public boolean remove(Integer employeeID) throws PersistenceException;
     public Employee getEmployeeByID(Integer employeeID) throws PersistenceException;
