+++ LeMaitreCore/src/br/cefetmg/LeMaitre/model/service/EmployeeManagemant.java
@@ -14,7 +14,7 @@ import br.cefetmg.LeMaitre.model.exception.PersistenceException;
  * @author Thalesgsn
  */
 public interface EmployeeManagemant {
-    public Long employeeInsert(Employee employee) throws BusinessException, PersistenceException;
+    public Integer employeeInsert(Employee employee) throws BusinessException, PersistenceException;
     public boolean employeeUpdate(Employee employee) throws BusinessException, PersistenceException;
     public boolean employeeRemove(Integer employeeID) throws PersistenceException;
     public Employee getEmployeeByID(Integer employeeID) throws PersistenceException;
